Cloud/ServerlessDesignPatternsandBestPractices_Code/Chapter02/requirements.txt
factory_boy
psycopg2
schematics
simplejson
sqlalchemy-utils
sqlalchemy>=1.2.0b2
